requirements.txt
Django==5.2.4
djangorestframework==3.15.2
celery==5.3.4
kombu==5.3.4
billiard==4.2.0
django-celery-results==2.5.1
pytz==2024.1
sqlparse==0.5.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
redis==5.0.1
requests==2.31.0
django-redis==5.4.0
django-filter==24.3
gunicorn==21.2.0
djangorestframework-simplejwt==5.5.1